+++ test/text/Paragraph.spec.ts
@@ -1,4 +1,4 @@
-import { HorizontalAlignment } from "../../src/style/HorizontalAlignment";
+import { join } from "path";
 import { Style } from "../../src/style/Style";
 import { Paragraph } from "../../src/text/Paragraph";
 import { TextDocument } from "../../src/TextDocument";
@@ -14,40 +14,33 @@ describe(Paragraph.name, () => {
     it("add text namespace", () => {
       document.addParagraph();
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/xmlns:text="urn:oasis:names:tc:opendocument:xmlns:text:1.0"/);
+      expect(document.toString()).toMatch(/xmlns:text="urn:oasis:names:tc:opendocument:xmlns:text:1.0"/);
     });
 
     it("insert an empty paragraph", () => {
       document.addParagraph();
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/<text:p\/>/);
+      expect(document.toString()).toMatch(/<text:p\/>/);
     });
 
     it("insert a paragraph with specified text", () => {
       document.addParagraph("some text");
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/<text:p>some text<\/text:p>/);
+      expect(document.toString()).toMatch(/<text:p>some text<\/text:p>/);
     });
   });
 
   describe("#addText", () => {
     it("set the text if element is empty", () => {
-      const paragraph = document.addParagraph();
-      paragraph.addText("some text");
+      document.addParagraph().addText("some text");
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/<text:p>some text<\/text:p>/);
+      expect(document.toString()).toMatch(/<text:p>some text<\/text:p>/);
     });
 
     it("append the text", () => {
-      const paragraph = document.addParagraph("some text");
-      paragraph.addText(" some more text");
+      document.addParagraph("some text").addText(" some more text");
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/<text:p>some text some more text<\/text:p>/);
+      expect(document.toString()).toMatch(/<text:p>some text some more text<\/text:p>/);
     });
   });
 
@@ -64,52 +57,54 @@ describe(Paragraph.name, () => {
 
   describe("#setText", () => {
     it("replace existing text with specified text", () => {
-      const paragraph = document.addParagraph("some text");
-      paragraph.setText("some other text");
+      document.addParagraph("some text").setText("some other text");
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/<text:p>some other text<\/text:p>/);
+      expect(document.toString()).toMatch(/<text:p>some other text<\/text:p>/);
     });
   });
 
   it("replace newline with line break", () => {
     document.addParagraph("some text\nsome more text");
 
-    const documentAsString = document.toString();
-    expect(documentAsString).toMatch(/<text:p>some text<text:line-break\/>some more text<\/text:p>/);
+    expect(document.toString()).toMatch(/<text:p>some text<text:line-break\/>some more text<\/text:p>/);
   });
 
   describe("#addHyperlink", () => {
     it("append a linked text", () => {
-      const paragraph = document.addParagraph("some text");
-      paragraph.addHyperlink(" some linked text", "http://example.org/");
+      document.addParagraph("some text").addHyperlink(" some linked text", "http://example.org/");
 
-      const documentAsString = document.toString();
       /* tslint:disable-next-line:max-line-length */
-      expect(documentAsString).toMatch(/<text:p>some text<text:a xlink:type="simple" xlink:href="http:\/\/example.org\/"> some linked text<\/text:a><\/text:p>/);
+      expect(document.toString()).toMatch(/<text:p>some text<text:a xlink:type="simple" xlink:href="http:\/\/example.org\/"> some linked text<\/text:a><\/text:p>/);
     });
+  });
 
-    it("not create a hyperlink if text is empty", () => {
-      const paragraph = document.addParagraph("some text");
-      paragraph.addHyperlink("", "http://example.org/");
+  describe("#addImage", () => {
+    it("append a draw frame with image and binary data", () => {
+      document.addParagraph().addImage(join(__dirname, "..", "data", "ODF.png"));
 
-      const documentAsString = document.toString();
-      expect(documentAsString).toMatch(/<text:p>some text<\/text:p>/);
+      const regex = new RegExp("<draw:frame text:anchor-type=\"paragraph\">"
+        + "<draw:image>"
+        + "<office:binary-data>"
+        + ".*"
+        + "</office:binary-data>"
+        + "</draw:image>"
+        + "</draw:frame>");
+      expect(document.toString()).toMatch(regex);
     });
   });
 
   describe("#setStyle", () => {
     let paragraph: Paragraph;
-    let style: Style;
+    let testStyle: Style;
 
     beforeEach(() => {
       paragraph = document.addParagraph("some text");
-      style = new Style();
+      testStyle = new Style();
     });
 
     it("set style-name attribute on paragraph if any style property was set", () => {
-      style.setPageBreakBefore();
-      paragraph.setStyle(style);
+      testStyle.setPageBreakBefore();
+      paragraph.setStyle(testStyle);
 
       expect(document.toString()).toMatch(/<text:p text:style-name="([a-z0-9]+)">some text<\/text:p>/);
     });
@@ -121,7 +116,7 @@ describe(Paragraph.name, () => {
     });
 
     it("not style-name attribute if default style is set", () => {
-      paragraph.setStyle(new Style());
+      paragraph.setStyle(testStyle);
 
       expect(document.toString()).toMatch(/<text:p>some text<\/text:p>/);
     });
@@ -140,7 +135,6 @@ describe(Paragraph.name, () => {
 
     it("return previous set style", () => {
       const testStyle = new Style();
-      testStyle.setHorizontalAlignment(HorizontalAlignment.Center);
       testStyle.setPageBreakBefore();
 
       paragraph.setStyle(testStyle);